state uplift of filter function from index.js to _app.js
